Static cache abstracting into the new functions

+++ classes/PodsView.php
@@ -202,71 +202,81 @@ class PodsView {
 			}
 		}
 
+		$cache_enabled = ! in_array( $cache_mode, $nocache, true );
+
 		if ( apply_filters( 'pods_view_cache_alt_get', false, $cache_mode, $group_key . $key, $original_key, $group ) ) {
 			$value = apply_filters( 'pods_view_cache_alt_get_value', $value, $cache_mode, $group_key . $key, $original_key, $group );
-		} elseif ( 'transient' === $cache_mode && ! in_array( $cache_mode, $nocache ) ) {
-			$value = get_transient( $group_key . $key );
-		} elseif ( 'site-transient' === $cache_mode && ! in_array( $cache_mode, $nocache ) ) {
-			$value = get_site_transient( $group_key . $key );
-		} elseif ( 'cache' === $cache_mode && $object_cache && ! in_array( $cache_mode, $nocache ) ) {
-			$value = wp_cache_get( $key, ( empty( $group ) ? 'pods_view' : $group ) );
-		} elseif ( 'option-cache' === $cache_mode && ! in_array( $cache_mode, $nocache ) ) {
-			global $_wp_using_ext_object_cache;
+		} elseif ( $cache_enabled ) {
+			if ( 'transient' === $cache_mode ) {
+				$value = get_transient( $group_key . $key );
+			} elseif ( 'site-transient' === $cache_mode ) {
+				$value = get_site_transient( $group_key . $key );
+			} elseif ( 'cache' === $cache_mode && $object_cache ) {
+				$value = wp_cache_get( $key, ( empty( $group ) ? 'pods_view' : $group ) );
+			} elseif ( 'option-cache' === $cache_mode ) {
+				global $_wp_using_ext_object_cache;
 
-			$pre = apply_filters( "pre_transient_{$key}", false );
+				$pre = apply_filters( "pre_transient_{$key}", false );
 
-			if ( false !== $pre ) {
-				$value = $pre;
-			} elseif ( $_wp_using_ext_object_cache ) {
-				$cache_found = false;
+				if ( false !== $pre ) {
+					$value = $pre;
+				} elseif ( $_wp_using_ext_object_cache ) {
+					$cache_found = false;
 
-				$value = wp_cache_get( $key, ( empty( $group ) ? 'pods_option_cache' : $group ), false, $cache_found );
+					$value = wp_cache_get( $key, ( empty( $group ) ? 'pods_option_cache' : $group ), false, $cache_found );
 
-				if ( false === $value || ! $cache_found ) {
-					if ( is_callable( $callback ) ) {
-						// Callback function should do it's own set/update for cache
-						$callback_value = call_user_func( $callback, $original_key, $group, $cache_mode );
+					if ( false === $value || ! $cache_found ) {
+						if ( is_callable( $callback ) ) {
+							// Callback function should do it's own set/update for cache
+							$callback_value = call_user_func( $callback, $original_key, $group, $cache_mode );
 
-						if ( null !== $callback_value && false !== $callback_value ) {
-							$value = $callback_value;
+							if ( null !== $callback_value && false !== $callback_value ) {
+								$value = $callback_value;
+							}
+
+							$called = true;
 						}
-
-						$called = true;
 					}
+				} else {
+					$transient_option  = '_pods_option_' . $key;
+					$transient_timeout = '_pods_option_timeout_' . $key;
+
+					$value   = get_option( $transient_option );
+					$timeout = get_option( $transient_timeout );
+
+					if ( ! empty( $timeout ) && $timeout < time() ) {
+						if ( is_callable( $callback ) ) {
+							// Callback function should do it's own set/update for cache
+							$callback_value = call_user_func( $callback, $original_key, $group, $cache_mode );
+
+							if ( null !== $callback_value && false !== $callback_value ) {
+								$value = $callback_value;
+							}
+
+							$called = true;
+						} else {
+							$value = false;
+
+							delete_option( $transient_option );
+							delete_option( $transient_timeout );
+						}
+					}
+				}//end if
+
+				if ( false !== $value ) {
+					$value = apply_filters( "transient_{$key}", $value );
+				}
+			} elseif ( 'static-cache' === $cache_mode ) {
+				$static_cache = pods_container( Static_Cache::class );
+
+				if ( $static_cache ) {
+					$value = $static_cache->get( $key, ( empty( $group ) ? 'pods_view' : $group ) );
+				} else {
+					$value = false;
 				}
 			} else {
-				$transient_option  = '_pods_option_' . $key;
-				$transient_timeout = '_pods_option_timeout_' . $key;
-
-				$value   = get_option( $transient_option );
-				$timeout = get_option( $transient_timeout );
-
-				if ( ! empty( $timeout ) && $timeout < time() ) {
-					if ( is_callable( $callback ) ) {
-						// Callback function should do it's own set/update for cache
-						$callback_value = call_user_func( $callback, $original_key, $group, $cache_mode );
-
-						if ( null !== $callback_value && false !== $callback_value ) {
-							$value = $callback_value;
-						}
-
-						$called = true;
-					} else {
-						$value = false;
-
-						delete_option( $transient_option );
-						delete_option( $transient_timeout );
-					}
-				}
+				$value = false;
 			}//end if
-
-			if ( false !== $value ) {
-				$value = apply_filters( "transient_{$key}", $value );
-			}
-		} elseif ( 'static-cache' === $cache_mode && ! in_array( $cache_mode, $nocache ) ) {
-			$static_cache = pods_container( Static_Cache::class );
-
-			$value = $static_cache->get( $key, ( empty( $group ) ? 'pods_view' : $group ) );
 		} else {
 			$value = false;
 		}//end if
@@ -367,7 +377,9 @@ class PodsView {
 		} elseif ( 'static-cache' === $cache_mode ) {
 			$static_cache = pods_container( Static_Cache::class );
 
-			$static_cache->set( $key, $value, ( empty( $group ) ? __CLASS__ : $group ) );
+			if ( $static_cache ) {
+				$static_cache->set( $key, $value, ( empty( $group ) ? __CLASS__ : $group ) );
+			}
 		}//end if
 
 		do_action( "pods_view_set_{$cache_mode}", $original_key, $value, $expires, $group );
@@ -476,10 +488,12 @@ class PodsView {
 		} elseif ( 'static-cache' === $cache_mode ) {
 			$static_cache = pods_container( Static_Cache::class );
 
-			if ( true === $key ) {
-				$static_cache->flush( ( empty( $group ) ? 'pods_view' : $group ) );
-			} else {
-				$static_cache->delete( ( empty( $key ) ? 'pods_view' : $key ), ( empty( $group ) ? 'pods_view' : $group ) );
+			if ( $static_cache ) {
+				if ( true === $key ) {
+					$static_cache->flush( ( empty( $group ) ? 'pods_view' : $group ) );
+				} else {
+					$static_cache->delete( ( empty( $key ) ? 'pods_view' : $key ), ( empty( $group ) ? 'pods_view' : $group ) );
+				}
 			}
 		}//end if
